Increase default servire startup timeout to 60 seconds
In the future, we'll need a way to make this more flexible and be able
to wait for more arbitrary and customizable conditions to declare a
service is "up". This could involve executing a script to get the ZK
ensemble status, waiting for the sending of a specific signal, etc.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>