Implement --only and understand services and containers interchangeably
Service names and container names can now be used (and mixed) on the
command line when passing a list of "things" to act on. If there is an
ambiguity in the naming, the smallest entity (the container) will be
considered as the safest choice.

Implement the -o/--only flag that restrict the action of the Maestro
score to only the container(s) and/or service(s) explicitely passed on
the command-line. Otherwise, dependencies or dependents (depending on
the action) are considered as well.

The shell completion helper was also modified accordingly, now returning
the set of all service and container names as possible values for the
third parameter.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>