Code refactoring and cleanup
Moved the entities definitions to entities.py to improve the readability
of the main Conductor class.

Created Maestro-specific exceptions into the exceptions.py module.
Correctly raise the relevant exceptions in the various parts of the
code, catch them in the main() function so we can correctly exit with an
error code when an exception occurs.

Removed useless debug logging and corresponding -v/--verbose option.

Style fixes for PEP-8 compliance.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>