Include a <service_name>_INSTANCE link variable
The <service_name>_INSTANCE variable is inserted into the environment
for each dependent service and contains a comma-separated list of the
instance names of all the containers that make up this service. It is
then used in the guestutils to get said list instead of relying on
parsing the <service_name>_(\w+)_HOST environment variable, which were
subject to fuzzy matching when two service names share the same prefix.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>