Fix restart orchestration play
Affected containers should be the same as for a stop (i.e. you first
stop dependents, then start everything back in start order, but for
dependents, not dependencies).

Display order is still incorrect, but orchestration actions are carried
out in the correct order because the parallelization checks force it to
be.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>