SSH tunneling support. Fixes #35 and #44.
Adds support for connecting to a remote Docker daemon via an SSH tunnel.
Uses bgtunnel to create an on-demand port forwarding tunnel to the
remote host (using the provided user and identity file). This allows for
the Docker daemons to only listen on localhost on the host they run on
(using `-H tcp://127.0.0.1:2375`) and have the SSH tunnel secure the
connection to them.

Signed-off-by: Maxime Petazzoni <max@signalfuse.com>