Fix a bug in error handling for Timeouts
When a Timeout is caught, and `msg = e.args[0][1]` errors and is handled,
the subsequent `raise` does not raise the Timeout as intended, but instead
raises the IndexError that was previously handled, as it was the most recent exception.

This is fixed by saving the exception and traceback info, then raising
it explicitly instead of raising the most recent exception.