WebHook auditor
This provides a simple webhook-based auditor that can make HTTP requests
to a given endpoint for audit. The webhook will make a request to the
configured endpoint with a JSON payload. The what, who, action and
message can be fed into the payload by using the corresponding
replacement strings {what}, {who}, {action} and {message}. See demo.yaml
for an example configuration.

The HTTP method and timeouts can also be configured.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>