Add support for omit: true. Fixes #108
Allow for services to not be automatically started when executing
non-specific actions with Maestro. Non-specific actions are commands
executed without specifying any service or container names as arguments.

This allows for services to be defined in the YAML, and used for
dependency definition purposes, but not actually acted upon unless it is
explicitly required, either by explicitly specifying it in the
arguments, or through a dependency requirement.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>