Display the correct image tag for running containers
If there are mismatches between the configuration and what's actually
running, make sure we always display the state of what is running. In
the case of the image tag, get it from the container status instead of
the container configuration when the container is running so we display
an accurate image tag.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>