Introduce --expand-services and --all on stop and restart commands
Maestro automatically expands "no arguments" to all the services in the
environment, and automatically expands a service name on the command
line to all the container instances of that service.

This is a useful behavior for most commands, but for `stop` and
`restart` commands it means that being "less specific" on the
command-line results in having a broader, potentially "more dangerous"
impact on the environment.

This change introduces two flags to help make human error in the usage
of those commands less impactful. The first one, `--expand-services`, is
required to expand a service name on the command line to its constituent
container instance names. For example:

  $ maestro -f myenv.yaml stop cassandra

Will no longer stop the entire `cassandra` service unless
`--expand-services` is also provided:

  $ maestro -f myenv.yaml stop --expand-services cassandra

The second one, `--all`, controls whether "no arguments" expands to all
services (it implies `--expand-services). To restart the whole
environment, one needs to write:

  $ maestro -f myenv.yaml restart --all

The other commands (`start`, `pull`, `clean`, `status`, ...) don't use
those new flags and will still expand and behave as before.