Add support for external environment files (#203)
Introduce support for external Docker environment files for services and
containers via the new `envfile` setting.

`envfile` may be a string, or list of strings, representing filenames
relative to the Maestro environment YAML file's location, to be used (in
order) for the environment of the service/container. The files are
loaded using docker-py's `parse_env_file()` helper.

The service or container environment passed via the existing `env`
setting is applied last and takes precedence over the values loaded from
the envfile(s).